// lib/api/generated/schemas/paginatedPublicShopsDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */
import type { PaginationMetaDto } from './paginationMetaDto';
import type { PublicShopListItemDto } from './publicShopListItemDto';

export interface PaginatedPublicShopsDto {
  data: PublicShopListItemDto[];
  meta: PaginationMetaDto;
}